Added completed challenge.

+++ Kattis/Python/TriangleArea.py
@@ -1,3 +1,3 @@
-height = float(input())
-base = float(input())
+height = int(input())
+base = int(input())
 print((0.5 * base) * height)
